requirements.txt
django-cms>=3.5,<3.6
djangocms-admin-style>=1.2,<1.3
django-treebeard>=4.0,<5.0

https://github.com/divio/djangocms-text-ckeditor/archive/master.zip?1521231604.2841148
https://github.com/divio/djangocms-file/archive/master.zip?1521231604.2841148
https://github.com/divio/djangocms-link/archive/master.zip?1521231604.2841148
https://github.com/divio/djangocms-style/archive/master.zip?1521231604.2841148
https://github.com/divio/djangocms-googlemap/archive/master.zip?1521231604.2841148
https://github.com/divio/djangocms-snippet/archive/master.zip?1521231604.2841148
https://github.com/divio/djangocms-picture/archive/master.zip?1521231604.2841148
https://github.com/divio/djangocms-video/archive/master.zip?1521231604.2841148
https://github.com/divio/djangocms-column/archive/master.zip?1521231604.2841148
easy_thumbnails
django-filer>=1.3
Django<2.0
pytz
django-classy-tags>=0.7
html5lib>=0.999999,<0.99999999
Pillow>=3.0
django-sekizai>=0.9
six